move metric enum one level up for public access.
